+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ### **(V.4.3.6 Changes) (1.18.1 Minecraft)**
 
 ##### Dimension:
-Fixed Bumblezone Dimension not using the random world seed in some edge cases. May produce chunk borders when loading into the dimension if created with past versions of Bumblezone.
+Fixed Bumblezone Dimension not using the random world seed in some edge cases. 
+ May produce chunk borders when loading into the dimension if created with past versions of Bumblezone.
 
 ##### Teleportation:
 Fixed teleportation getting screwed up when riding mobs and exiting Bumblezone. It should now use the controlling entity for destination instead of vehicle's data.